Bugfixes and support for individual stream processing
