+++ backend/package.json
@@ -31,6 +31,7 @@
     "format": "npx prettier --write .",
     "tsc-check": "tsc --noEmit",
     "script:process-integration": "SERVICE=script ts-node ./src/bin/scripts/process-integration.ts",
+    "script:process-stream": "SERVICE=script ts-node ./src/bin/scripts/process-stream.ts",
     "script:continue-run": "SERVICE=script ts-node ./src/bin/scripts/continue-run.ts",
     "script:change-tenant-plan": "SERVICE=script ts-node ./src/bin/scripts/change-tenant-plan.ts",
     "script:process-webhook": "SERVICE=script ts-node ./src/bin/scripts/process-webhook.ts",
